Move config update logic into pkg (just enough to compile and test ok)
