+++ pkg/config/error.go
@@ -1,0 +1,29 @@
+package config
+
+import "github.com/giantswarm/microerror"
+
+// executionFailedError should never be matched against and therefore there is
+// no matcher implement. For further information see:
+//
+//	https://github.com/giantswarm/fmt/blob/master/go/errors.md#matching-errors
+var executionFailedError = &microerror.Error{
+	Kind: "executionFailedError",
+}
+
+var invalidConfigError = &microerror.Error{
+	Kind: "invalidConfigError",
+}
+
+// IsInvalidConfig asserts invalidConfigError.
+func IsInvalidConfig(err error) bool {
+	return microerror.Cause(err) == invalidConfigError
+}
+
+var invalidFlagError = &microerror.Error{
+	Kind: "invalidFlagError",
+}
+
+// IsInvalidFlag asserts invalidFlagError.
+func IsInvalidFlag(err error) bool {
+	return microerror.Cause(err) == invalidFlagError
+}
